Merge netcdf files on the time dimension

This function merges all the netcdf files in a given directory over the time dimension. It saves the merged file in the working directory.

catNetcdf(inDir = NULL, varname = NULL, pattern = NULL,
  recursive = FALSE, outFileName = "outfile.nc", outDir = getwd())

Arguments

inDir
is the directory where all the files to read are stored

varname
name of the variable to extract

pattern
regular expression pattern to select a subset of files

recursive
logical (TRUE by default). If set to TRUE it looks in folders and subfolders

outFileName
output filename

outDir
is the directory where outFileName is saved, by default this is the working directory.

Examples

not_run({
    # Mergetime using single variable nc files
    catNetcdf(inDir = "/var/tmp/moc0/forestfire",
              varname = NULL,
              pattern = "geff_reanalysis_an_fwis_fwi_",
              recursive = TRUE,
              outFileName = "outfile.nc",
              outDir = getwd())

})