Generates the percentile risk index from a raster stack and a set of polygons

generates the mean of the values over a certain percentile theshold for the pixel of the raster stack under the polygons of a shapefile

getPercRiskIndex(r_stack, p_shape, perc.val = 75, mod = "gt")

Arguments

r_stack
is the raster or raster stack

p_shape
is the shapefile on which to aggregate the values

perc.val
is the percentile value used as a threshold

mod
defines if the values considered for the mean are above (gt) or below (lt) the threshold

Examples

not_run({
    r.index <- getPercRiskIndex(r_stack = r_stack, p_shape = poly,
                                perc.val = 50, mod = "lt")
})