get_fire_danger_levels

This function calculates the danger levels (VeryLow-Low-Moderate-High-VeryHigh-Extreme) for a given country.

get_fire_danger_levels(fire_index, ndays = 4)

Arguments

fire_index
RasterBrick containing the fire index to calculate the thresholds for. Please note that names(fire_index) should contain dates.

ndays
Number of days per year in which a fire is expected to occur. By default this is 4 days.

Value

A numeric vector listing the thresholds.

Examples

if (FALSE) {

  # Define period for Reanalysis
  dataDates <- seq.Date(from = as.Date(strptime(paste("1980", 1),
                                                format="%Y %j")),
                        to = as.Date(strptime(paste("2016", 366),
                                                format="%Y %j")),
                        by = "day")

  # Create an index of fire season dates
  seasons <- get_fire_season(dataDates, emisphere = "north")
  fireSeasonIndex <- which(seasons == TRUE)

  r <- raster::brick('FWI_1980-2016.nc')
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)
  maskcrop(r, p, mask = TRUE, crop = TRUE)

  # Mask and crop over country
  FWIcountry <- maskcrop(FWI, country, mask = TRUE, crop = TRUE)

  # Subset based on fire season
  FWIseasonal <- raster::subset(FWIcountry, fireSeasonIndex)

  # Generate thresholds for the country of interest
  countryThr <- get_fire_danger_levels(fire_index = FWIseasonal)

}