mask_with_fuelmodel

This function masks a Raster* object using JRC's fuelmodel.

mask_with_fuelmodel(x, fuelmap = NULL, codes = NULL, value = NA)

Arguments

x

Raster* to mask

fuelmap

Custom fuel map as RasterLayer.

codes

The value of the cells in fuelmap used to mask x.

value

The value of cells masked out by the fuel map (NA by default).

Details

In absence of vegetation the risk of ignition reduces considerably, regardless of the state of the soil. These areas are mapped in the JRC's fuelmodel (in use in 2018). This package contains a cached version of this map (matching crs and resolution of ERA-Interim), stored in the 'inst/extdata' folder of the package. For evaluating fire danger we use this map to mask deserts, glaciers, urban areas, etc. These areas have codes > 20. The optional argument `fuelmap` allows to input a custom fuelmap and costum `codes` to use for masking. Please note that

Examples

if (FALSE) {
  xmasked <- mask_with_fuelmodel(x)
}